% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomonic-main.R
\name{print.gnomos}
\alias{print.gnomos}
\title{Print method for gnomos class}
\usage{
\method{print}{gnomos}(x, ...)
}
\arguments{
\item{x}{an object class 'gnomos'.}

\item{\dots}{Additional arguments to the print method.}
}
\value{
The values of the proportionality constant (a), constant proportion of the overall natural death rate (G) and
a data.frame with the duration and natural mortality for each gnomonic interval.
#' @examples
model <- gnomonic(nInterval = 7, eggDuration = 2, addInfo = NULL,
longevity = 365, fecundity = 200000, a_init = 2)

print(model)
}
\description{
Print method for gnomos class
}
