% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomonicStochastic.R
\name{print.gnomosBoot}
\alias{print.gnomosBoot}
\title{Print method for gnomosBoot class}
\usage{
\method{print}{gnomosBoot}(x, ...)
}
\arguments{
\item{x}{an object class 'gnomosBoot'.}

\item{\dots}{Additional arguments to the print method.}
}
\value{
The values of the proportionality constant (a), constant proportion of the overall natural death rate (G) and
a data.frame with the duration and natural mortality for each gnomonic interval.
}
\description{
Print method for gnomosBoot class
}
\examples{
#The values are based on Caddy (1996).
modelBoot <- gnomonicStochastic(nInterval = 7, eggDuration = 2, addInfo = NULL, longevity = 365,
distr = "uniform", min_fecundity = 100000, max_fecundity = 300000, niter = 50, a_init = 2)

print(modelBoot)
}
